+++ test/test_main.py
@@ -32,7 +32,7 @@ class HighlightTestCase(unittest.TestCase):
 
         self.assertIsNone(err_msg)
         self.assertEqual(
-            '\'<code class="gch-pygments">'
+            '\'<code class="gch-pygments"><!-- gch-lang: python -->'
             + '<span class="mi">123</span>'
             + "</code>'",
             editor.unwrap_action.contents,
@@ -52,7 +52,7 @@ class HighlightSelectionTestCase(unittest.TestCase):
         )
 
         self.assertEqual(
-            '<code class="gch-pygments">'
+            '<code class="gch-pygments"><!-- gch-lang: python -->'
             + '<span class="k">return</span> <span class="mi">123</span>'
             + "</code>",
             str(result),
@@ -79,6 +79,8 @@ class HighlightSelectionTestCase(unittest.TestCase):
         )
 
         self.assertEqual(
-            '<code class="gch-pygments">' + '<span class="mi">123</span>' + "</code>",
+            '<code class="gch-pygments"><!-- gch-lang: python -->'
+            + '<span class="mi">123</span>'
+            + "</code>",
             str(result),
         )