+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog][kac], and this project adheres to
 ### Added
 
 - A fuzzy finder for language selection.
+- Language comment to highlighted code for future migrations.
 
 ### Changed